+++ source/Runtime.pas
@@ -243,19 +243,12 @@ begin
 
     fImp.Free;
 
-
-    foAPI_Output.WriteLog('Start unload plugins');
-
     foPlugins.UnLoadPlugins;
-    foAPI_Output.WriteLog('End unload plugins');
-
 
     foAPI_Output.WriteLog('Logging finished');
     foAPI_Output.CloseLog;
 
     foAPI_Output.Free;
-
-
 
     foPlugins.Free;